Move EKS jobs cleanup to cleanup_k8s.py script
https://progress.opensuse.org/issues/124664
