+++ requirements_k8s.txt
@@ -1,3 +1,4 @@
 oauth2client
 kubernetes
 google-api-python-client==2.55.0
+boto3
