+++ tests/test_ec2.py
@@ -1,5 +1,3 @@
-import os
-import kubernetes
 from ocw.lib.EC2 import EC2, Provider
 from webui.PCWConfig import PCWConfig
 from tests.generators import mock_get_feature_property
@@ -445,140 +443,3 @@ def test_cleanup_all_calling_all(ec2_patch, monkeypatch):
     ec2_patch.cleanup_all()
 
     assert called_stack == ['cleanup_images', 'cleanup_snapshots', 'cleanup_volumes', 'cleanup_uploader_vpcs']
-
-
-def test_list_clusters(ec2_patch, monkeypatch):
-    mocked_eks = MockedEKSClient()
-    monkeypatch.setattr(EC2, 'eks_client', lambda self, region: mocked_eks)
-    all_clusters = ec2_patch.all_clusters()
-    assert all_clusters == {}
-
-    mocked_eks.clusters_list = {'clusters': ['empty']}
-    all_clusters = ec2_patch.all_clusters()
-    assert all_clusters == {}
-
-    mocked_eks.clusters_list = {'clusters': ['hascluster']}
-    all_clusters = ec2_patch.all_clusters()
-    assert all_clusters == {}
-
-    mocked_eks.clusters_list = {'clusters': ['hastags']}
-    all_clusters = ec2_patch.all_clusters()
-    assert all_clusters == {'region1': ['hastags']}
-
-    mocked_eks.clusters_list = {'clusters': ['hastags', 'ignored']}
-    all_clusters = ec2_patch.all_clusters()
-    assert all_clusters == {'region1': ['hastags']}
-
-
-class MockedKubernetesConfig():
-    def load_kube_config(self, *args, **kwargs):
-        return True
-
-
-class MockedKubernetesClient():
-    def __init__(self, jobs=[]):
-        self.jobs = jobs
-        self.deleted_jobs = []
-
-    # pylint: disable=C0103
-    def BatchV1Api(self):
-        return self
-
-    def list_job_for_all_namespaces(self, *args, **kwargs):
-        return MockedKubernetesResult(self.jobs)
-
-    def delete_namespaced_job(self, name, namespace):
-        self.deleted_jobs.append(name)
-
-
-class MockedKubernetesResult():
-    def __init__(self, items):
-        self.items = items
-
-
-class MockedKubernetesJobStatus():
-    def __init__(self, age):
-        self.start_time = datetime.now(timezone.utc) - timedelta(days=age)
-
-
-class MockedKubernetesJobMetadata():
-    def __init__(self, name):
-        self.name = name
-        self.namespace = "default"
-
-
-class MockedKubernetesJob():
-    def __init__(self, name, age):
-        self.status = MockedKubernetesJobStatus(age)
-        self.metadata = MockedKubernetesJobMetadata(name)
-
-
-class MockedSubprocessReturn():
-    def __init__(self, returncode=0, stderr=""):
-        self.returncode = returncode
-        self.stderr = stderr
-
-
-@pytest.fixture
-def k8s_patch(monkeypatch):
-    monkeypatch.setattr(EC2, 'check_credentials', lambda *args, **kwargs: True)
-    monkeypatch.setattr(Provider, 'read_auth_json', lambda *args, **kwargs: '{}')
-    monkeypatch.setattr(EC2, 'get_all_regions', lambda self: ['region1'])
-    monkeypatch.setattr(PCWConfig, 'get_feature_property', mock_get_feature_property)
-
-    return EC2('fake')
-
-
-def test_kubectl_client(k8s_patch, monkeypatch):
-    monkeypatch.setattr(Provider, 'cmd_exec', lambda *args, **kwargs: MockedSubprocessReturn(0))
-    monkeypatch.setattr(kubernetes, 'config', MockedKubernetesConfig())
-    mocked_client1 = MockedKubernetesClient(1)
-    monkeypatch.setattr(kubernetes, 'client', mocked_client1)
-    assert k8s_patch.kubectl_client("region1", "cluster") == mocked_client1
-
-    # Check that the client is reused
-    mocked_client2 = MockedKubernetesClient(1)
-    monkeypatch.setattr(kubernetes, 'client', mocked_client2)
-    assert k8s_patch.kubectl_client("region1", "cluster") == mocked_client1
-
-    # Invalid 'aws eks update-kubeconfig' execution should return None
-    monkeypatch.setattr(Provider, 'cmd_exec', lambda *args, **kwargs: MockedSubprocessReturn(1))
-    assert k8s_patch.kubectl_client("region2", "cluster") is None
-
-
-def test_create_credentials_file(k8s_patch, monkeypatch):
-    monkeypatch.setattr(Provider, 'cmd_exec', lambda *args, **kwargs: MockedSubprocessReturn(0))
-    k8s_patch.create_credentials_file("/tmp")
-    assert os.path.exists("/tmp/.aws/credentials")
-
-    # Invalid credentials, 'aws sts get-caller-identity' fails
-    monkeypatch.setattr(Provider, 'cmd_exec', lambda *args, **kwargs: MockedSubprocessReturn(1, "test"))
-    error = None
-    try:
-        k8s_patch.create_credentials_file("/tmp")
-    except Exception as exception:
-        error = exception
-
-    assert error is not None
-    assert str(error) == "Invalid credentials, the credentials cannot be verified by'aws " + \
-                         "sts get-caller-identity' with the error: test"
-
-
-def test_cleanup_k8s_jobs(k8s_patch, monkeypatch):
-    mocked_eks = MockedEKSClient()
-    mocked_eks.clusters_list = {'clusters': ['cluster1']}
-    monkeypatch.setattr(EC2, 'eks_client', lambda self, region: mocked_eks)
-
-    monkeypatch.setattr(EC2, 'create_credentials_file', lambda *args, **kwargs: True)
-    monkeypatch.setattr(kubernetes, 'config', MockedKubernetesConfig())
-    mocked_kubernetes = MockedKubernetesClient([MockedKubernetesJob("job1", 1), MockedKubernetesJob("job2", 0)])
-    monkeypatch.setattr(EC2, "kubectl_client", lambda *args, **kwargs: mocked_kubernetes)
-    k8s_patch.cleanup_k8s_jobs()
-    assert len(mocked_kubernetes.deleted_jobs) == 1
-    assert mocked_kubernetes.deleted_jobs[0] == "job1"
-
-    # test dry_run
-    k8s_patch.dry_run = True
-    mocked_kubernetes.deleted_jobs = []
-    k8s_patch.cleanup_k8s_jobs()
-    assert len(mocked_kubernetes.deleted_jobs) == 0
